+++ dataDisplay/flaskapp/sums_models/update.py
@@ -1,60 +1,84 @@
 # coding=utf-8
-from dataDisplay.flaskapp.sums_models.model_sums import *
+from dataDisplay.flaskapp.sums_models.methods import *
+from dataDisplay.flaskapp.rules.methods import *
+import datetime
 
-def delet_data(table_name, id):
+log_path = 'datadisplay/flaskapp/data/source_logs.txt'
+log_path2 = 'datadisplay/flaskapp/data/sums_logs.txt'
+
+def update_db(path, xls_name):
+
+    # 待更新excle文件，同步到数据库中
+    # 获取文件中的所有的sheets，以及对应数据库中的英文表名
+    all_datas, sheet_names = give_sheet(path, xls_name)
+    ref_names = page_tables()
+    columns = show_columns()
+
+    time = datetime.datetime.now().strftime('%Y-%m-%d %H:%M:%S')
+    logs = open(log_path, 'a')
+    logs.write(time + ': 科室数据更新结果\n')
+
+    for sheet in sheet_names:
+        logs.write('sheet:' + sheet + '\n')
+
+        datas = all_datas[sheet]
+        ref_name = ref_names[sheet]
+        column = columns[ref_name][1]
+
+        for e in datas:
+            # 逐条读取单个sheet中的数据，要重新构建成二位list，直接用list函数不行
+            data = []
+            data.append(e)
+
+            fail_lists = []
+            fail_sums = 0
+            try:
+                insert(ref_name, data, column)
+            except:
+                id = data[0]
+                try:
+                    update_data(ref_name, id, data, column)
+                except:
+                    fail_lists.append(id)
+                    fail_sums += 1
+
+        logs.write('数据上传结束！\n')
+        if fail_sums:
+            logs.write('有部分数据未能完成更新，序号为：' + ','.join(fail_lists) + '\n')
+        else:
+            logs.write('所有数据更新完成！\n')
+
+def update_sums(table_name, sums, columns):
     '''
-    删除对应id行的数据
+    更新数据汇总表
     :param table_name:
-    :param id:
-    :return:
-    '''
-    exec(table_name+ '.query.filter_by(id = id).delete()')
-    db.session.commit()
-
-def update_data(tale_name, id, new_data, columns):
-    '''
-    更新表中的某行数据, 以id（主键）为索引目标
-    :param tale_name:
-    :param id:
-    :param new_data:
+    :param sums:
     :param columns:
     :return:
     '''
-    values = []
-    for index in range(len(new_data)):
-        # 如果new_Data[index]中包含换行符，会报错, 需要保持字符串为raw string（repr, eval）
-        # 接受更新值长度比原栏目长度短的情况
-        data = str(new_data[index]).replace('\n', ' ')
-        value = tale_name + '.' + columns[index] + ':' + "'" + data + "'"
-        values.append(value)
+    logs = open(log_path2, 'a')
+    time = datetime.datetime.now().strftime('%Y-%m-%d %H:%M:%S')
+    logs.write(time + ': 汇总表自动统计数据更新\n')
 
-    values = '{' + ', '.join(values) + '}'
-    print 'values', values
+    for e in sums:
+        # 逐条读取单个sheet中的数据，要重新构建成二位list，直接用list函数不行
+        data = []
+        data.append(e)
 
-    exec( tale_name+'.query.filter_by(id = id).update(' + values + ')')
-    db.session.commit()
-
-def insert(table_name, xls_data, columns):
-    '''
-    全表格查找，匹配表格中的所有栏目
-    :param table_name:
-    :param xls_data:
-    :param columns:
-    :return:
-    '''
-    for data in xls_data:
-        content = None
-
+        fail_lists = []
+        fail_sums = 0
         try:
-            exec('content ='+ table_name + '(columns, data)')
-            db.session.add(content)
-        except Exception, e:
-            # models没有成功初始化
-            print 'table model init error: ', table_name, len(columns), len(data)
-            print e
+            insert(table_name, data, columns)
+        except:
+            id = data[0]
+            try:
+                update_data(table_name, id, data, columns)
+            except:
+                fail_lists.append(id)
+                fail_sums += 1
 
-        try:
-            db.session.commit()
-        except Exception, e:
-            # print 'ERROR:', e
-            db.session.rollback()
+    logs.write('数据上传结束！\n')
+    if fail_sums:
+        logs.write('有部分数据未能完成更新，序号为：' + ','.join(fail_lists) + '\n')
+    else:
+        logs.write('所有数据更新完成！\n')
